Decode HTML entities for previews and search index
fixes #15

we use the "he" library for this as the stripping has to happen at build
time in `extendPageData`, i.e. in a NodeJS process that does not have
access to the browser facilities for converting HTML entities.

We don't just strip the entities as depending on the language, they may
be important searchable characters (e.g. special non-ASCII letters of
the language).